Usability improvements
Monte Carlo simulation now allows user to specify the folder they wish results to go into. If the folder does not exist, it is created
The data plotting function now takes in a an array of strings to use as descriptors for the graphs so that the user does not have to manually input them if they draw a legend.