Multithreaded Euler Characteristic
The Euler Characteristic can also be expressed as a sum of the local curvature at every node. The Euler Characteristic algorithm has been rewritten to exploit this and allow multithreading. Using the curvatureAt function will store the value for later use so it does not have to be recalculated by the Euler characteristic (or further calls to curvatureAt).